Phase 2: Screen capture + raw display pipeline
Add X11 XShm screen capture, SDL3 YUV420p renderer, and session
orchestrators for uncompressed screen sharing over LAN at ~10fps.

- Fix protocol header: sequence/frag_idx/frag_total to uint16_t,
  supporting >255 fragments per frame (needed for raw 1080p YUV420p)
- Bump PROTOCOL_VERSION to 2
- Add ICaptureSource interface and ScreenCaptureX11 (XShm + swscale)
- Add SdlRenderer with IYUV streaming texture
- Add HostSession (capture/send/poll threads) and ClientSession
  (recv thread + main-thread SDL render loop)
- Replace synthetic test data in main.cpp with real sessions
- Add --fps flag (default 10)
- Add test_large_frame_roundtrip (3.1MB frame, out-of-order, wraparound)

+++ src/net/packet_assembler.h
@@ -20,8 +20,8 @@ public:
 private:
     struct FrameState {
         uint16_t frame_id = 0;
-        uint8_t frag_total = 0;
-        uint8_t frags_received = 0;
+        uint16_t frag_total = 0;
+        uint16_t frags_received = 0;
         PacketType type = PacketType::VIDEO_DATA;
         uint8_t flags = 0;
         uint32_t timestamp_us = 0;
