+++ src/net/packet_fragmenter.h
@@ -11,7 +11,7 @@ class PacketFragmenter {
 public:
     // Fragments an encoded packet into UDP-sized Packets.
     // Each fragment gets a 16-byte header + up to MAX_FRAGMENT_DATA bytes of payload.
-    std::vector<Packet> fragment(const EncodedPacket& encoded, uint32_t& sequence);
+    std::vector<Packet> fragment(const EncodedPacket& encoded, uint16_t& sequence);
 };
 
 } // namespace lancast
